+++ gcovr/writer/summary.py
@@ -27,15 +27,19 @@ def print_summary(covdata):
     '''
 
     (lines_total, lines_covered, percent,
-     branches_total, branches_covered,
-     percent_branches) = get_global_stats(covdata)
+     functions_total, functions_covered, percent_functions,
+     branches_total, branches_covered, percent_branches) = get_global_stats(covdata)
 
     lines_out = "lines: %0.1f%% (%s out of %s)\n" % (
         percent, lines_covered, lines_total
+    )
+    functions_out = "functions: %0.1f%% (%s out of %s)\n" % (
+        percent_functions, functions_covered, functions_total
     )
     branches_out = "branches: %0.1f%% (%s out of %s)\n" % (
         percent_branches, branches_covered, branches_total
     )
 
     sys.stdout.write(lines_out)
+    sys.stdout.write(functions_out)
     sys.stdout.write(branches_out)